% Generated by build script
\name{msci_monthly}
\alias{msci_monthly}
\title{Fetch MSCI index monthly level data}
\usage{
msci_monthly()
}
\arguments{
\item{codes}{Character vector of MSCI index codes}
\item{start}{Start date in "YYYYMMDD" format (default: 5 years ago)}
\item{end}{End date in "YYYYMMDD" format (default: today)}
\item{variant}{Index variant: "STRD", "NETR", or "GRTR". Default "STRD".}
\item{currency}{Currency code (default "USD")}
}
\value{tibble: date, level, index_code, index_variant, currency}
\description{Fetch MSCI index monthly level data}
